add auto mutation and manual mutation

+++ projects/client/src/pages/admin/ManageWarehouse.jsx
@@ -76,25 +76,37 @@ const ManageWarehouse = () => {
   const deleteWarehouse = async (id) => {
     // console.log(id);
     // let wId = id;
-    try {
-      const response = await api.delete(`/warehouses/delete/${id}`);
-      console.log(response);
-      getWarehouseData();
-      Swal.fire({
-        title: "Success",
-        text: response.data.message,
-        icon: "success",
-        confirmButtonText: "Ok",
-      });
-    } catch (error) {
-      Swal.fire({
-        title: "Error!",
-        text: error.response.data.message,
-        icon: "warning",
-        confirmButtonText: "Ok",
-      });
-      console.log(error);
-    }
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const response = await api.delete(`/warehouses/delete/${id}`);
+          console.log(response);
+          getWarehouseData();
+          Swal.fire({
+            title: "Success",
+            text: response.data.message,
+            icon: "success",
+            confirmButtonText: "Ok",
+          });
+        } catch (error) {
+          Swal.fire({
+            title: "Error!",
+            text: error.response.data.message,
+            icon: "warning",
+            confirmButtonText: "Ok",
+          });
+          console.log(error);
+        }
+      }
+    });
   };
 
   useEffect(() => {
@@ -120,7 +132,12 @@ const ManageWarehouse = () => {
           {el.city}
         </td>
         <td className="whitespace-nowrap px-2 py-2 text-sm text-gray-500">
-          <Stack direction="row" spacing={0}>
+          <Stack
+            direction="row"
+            spacing={0}
+            display="flex"
+            alignContent="center"
+          >
             <EditeWarehouse wId={el.id} runFunction={getWarehouseData} />
             <Button
               variant="link"
@@ -138,12 +155,8 @@ const ManageWarehouse = () => {
 
   return (
     <div className={` ${paddingLeft}  py-10 items-center`}>
-      <h1>ini manage rpoduct</h1>
+      <h1>ini manage </h1>
       <div>
-        {/* <Button leftIcon={<AddIcon />}>Add Warehouse</Button> */}
-        {/* <Button leftIcon={<AddIcon />} onClick={onOpen}>
-          Add Warehouse
-        </Button> */}
         <AddWarehouse runFunction={getWarehouseData} />
       </div>
       <div className="mt-6 flex flex-col justify-end max-w-5xl xl">
@@ -181,65 +194,19 @@ const ManageWarehouse = () => {
                     <th
                       scope="col"
                       className="whitespace-nowrap px-2 py-3.5 text-left text-sm font-semibold text-gray-900"
-                    ></th>
+                    >
+                      Action
+                    </th>
                   </tr>
                 </thead>
                 <tbody className="divide-y divide-gray-200 bg-white">
                   {warehouse}
-                  {/* {transactions.map((transaction) => (
-                    <tr key={transaction.id}>
-                      <td className="whitespace-nowrap py-2 pl-4 pr-3 text-sm text-gray-500 sm:pl-6">
-                        {transaction.invoice_number.substr(0, 13)}
-                      </td>
-
-                      <td className="whitespace-nowrap px-2 py-2 text-sm text-gray-900">
-                        bgbgb
-                      </td>
-
-                      <td className="whitespace-nowrap px-2 py-2 text-sm text-gray-500">
-                        Rp. {transaction.total_price.toLocaleString("id-ID")}
-                      </td>
-                      <td className="whitespace-nowrap px-2 py-2 text-sm text-gray-500">
-                        <a
-                          href="#"
-                          className="text-indigo-600 hover:text-indigo-900"
-                        >
-                          Detail
-                        </a>
-                      </td>
-                      <td className="whitespace-nowrap px-2 py-2 text-sm text-gray-500">
-                        {transaction.status}
-                      </td>
-                      <td className="whitespace-nowrap py-2 pl-3 pr-4 text-right text-sm font-medium sm:pr-6">
-                        <button className="text-white rounded-md bg-black px-3 py-1 text-xs mr-4 hover:bg-gray-800">
-                          Confirm
-                        </button>
-                        <button className="text-black hover:text-indigo-900">
-                          Reject
-                        </button>
-                      </td>
-                    </tr>
-                  ))} */}
                 </tbody>
               </table>
             </div>
           </div>
         </div>
       </div>
-      {/* <TableContainer>
-        <Table size="sm">
-          <Thead>
-            <Tr>
-              <Th>No</Th>
-              <Th>Warehouse Name</Th>
-              <Th>Province</Th>
-              <Th>City</Th>
-              <Th>Action</Th>
-            </Tr>
-          </Thead>
-          <Tbody>{warehouse}</Tbody>
-        </Table>
-      </TableContainer> */}
 
       <Modal isOpen={isOpen} onClose={onClose}>
         <ModalOverlay />
